# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  app.CreditRequestInput:
    properties:
      amount:
        type: number
      memo:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - amount
    - memo
    type: object
  app.CreditUserResponse:
    properties:
      amount:
        type: number
      memo:
        type: string
      user:
        type: string
    type: object
...
  app.GenerateVouchersInput:
    properties:
      count:
        type: integer
      expire_after_days:
        type: integer
      value:
        type: number
    required:
    - count
    - expire_after_days
    - value
    type: object
...
  app.PendingRecordsResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      tft_amount:
        description: TFTs are multiplied by 1e7
        type: integer
      transferred_tft_amount:
        type: integer
      transferred_usd_amount:
        type: number
      updated_at:
        type: string
      usd_amount:
        type: number
      user_id:
        type: integer
    type: object
...
  models.Invoice:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/models.NodeItem'
        type: array
      tax:
        description: 'TODO:'
        type: number
      total:
        type: number
      user_id:
        type: integer
    required:
    - user_id
    type: object
  models.NodeItem:
    properties:
      contract_id:
        type: integer
      cost:
        description: in millicent
        type: number
      id:
        type: integer
      invoice_id:
        type: integer
      node_id:
        type: integer
      period:
        type: number
      rent_created_at:
        type: string
    type: object
...
  models.User:
    properties:
      account_address:
        type: string
      admin:
        type: boolean
      code:
        type: integer
      credit_card_balance:
        description: millicent, money from credit card
        type: integer
      credited_balance:
        description: millicent, manually added by admin or from vouchers
        type: integer
      debt:
        description: millicent
        type: integer
      email:
        type: string
      id:
        type: integer
      password:
        items:
          type: integer
        type: array
      sponsored:
        type: boolean
      ssh_key:
        type: string
      stripe_customer_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
      verified:
        type: boolean
    required:
    - email
    - password
    - username
    type: object
  models.Voucher:
    properties:
      code:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      redeemed:
        type: boolean
      value:
        type: number
    type: object
info:
  contact: {}
  description: API documentation for KubeCloud.
  title: KubeCloud API
  version: "1.0"
paths:
  /invoices:
    get:
      consumes:
      - application/json
      description: Returns a list of all invoices
      operationId: get-all-invoices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Invoice'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Get all invoices
      tags:
      - admin
...
  /pending-records:
    get:
      consumes:
      - application/json
      description: Returns all pending records in the system
      operationId: list-pending-records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.PendingRecordsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: List pending records
      tags:
      - admin
...
  /user/invoice:
    get:
      consumes:
      - application/json
      description: Returns a list of invoices for a user
      operationId: get-invoices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Invoice'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Get invoices
      tags:
      - invoices
  /user/invoice/{invoice_id}:
    get:
      consumes:
      - application/json
      description: Downloads an invoice by ID
      operationId: download-invoice
      parameters:
      - description: Invoice ID
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Invoice is not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Download invoice
      tags:
      - invoices
...
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of all users
      operationId: get-all-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Get all users
      tags:
      - admin
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user from the system
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/app.APIResponse'
        "403":
          description: Admins cannot delete their own account
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Delete a user
      tags:
      - admin
  /users/{user_id}/credit:
    post:
      consumes:
      - application/json
      description: Credits a specific user's balance
      operationId: credit-user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Credit Request Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.CreditRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app.CreditUserResponse'
        "400":
          description: Invalid request format or user ID
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Credit user balance
      tags:
      - admin
  /vouchers:
    get:
      consumes:
      - application/json
      description: Returns all vouchers in the system
      operationId: list-vouchers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Voucher'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: List vouchers
      tags:
      - admin
  /vouchers/generate:
    post:
      consumes:
      - application/json
      description: Generates a bulk of vouchers
      operationId: generate-vouchers
      parameters:
      - description: Generate Vouchers Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.GenerateVouchersInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Voucher'
            type: array
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Generate vouchers
      tags:
      - admin
swagger: "2.0"